+++ R/simStateSpace-ssm-intercept-eta.R
@@ -1,27 +1,23 @@
+#' Intercept from
 #' Steady-State Mean Vector for the
-#' Observed Variables in the
+#' Latent Variables in the
 #' State Space Model
 #'
-#' The steady-state mean vector
-#' for the observed variables
+#' The intercept vector
+#' for the latent variables
 #' in the state space model
-#' \eqn{\mathrm{Mean} \left( \mathbf{y} \right)}
+#' \eqn{\boldsymbol{\alpha}}
 #' is given by
 #' \deqn{
-#'   \mathrm{Mean} \left( \mathbf{y} \right)
+#'   \boldsymbol{\alpha}
 #'   =
-#'   \boldsymbol{\nu}
-#'   +
-#'   \boldsymbol{\Lambda}
-#'   \mathrm{Mean} \left( \boldsymbol{\eta} \right)
+#'   \mathrm{Mean} \left( \boldsymbol{\eta} \right) -
+#'   \boldsymbol{\beta} \mathrm{Mean} \left( \boldsymbol{\eta} \right)
 #' }
 #' where
-#' \eqn{\boldsymbol{\nu}}
-#' is the vector of intercept values
-#' for the measurement model,
-#' \eqn{\boldsymbol{\Lambda}}
-#' is the matrix of factor loadings,
-#' and
+#' \eqn{\boldsymbol{\beta}}
+#' is the transition matrix relating the values of the latent variables
+#' at the previous to the current time point,
 #' \eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}
 #' is the steady-state mean vector
 #' for the latent variables.
@@ -41,25 +37,21 @@
 #'   nrow = 3
 #' )
 #' alpha <- rep(x = 1, times = 3)
-#' lambda <- diag(3)
-#' nu <- rep(x = 1, times = 3)
 #' mean_eta <- SSMMeanEta(
 #'   beta = beta,
 #'   alpha = alpha
 #' )
-#' SSMMeanY(
-#'   nu = nu,
-#'   lambda = lambda,
+#' SSMInterceptEta(
+#'   beta = beta,
 #'   mean_eta = mean_eta
 #' )
 #'
 #' @family Simulation of State Space Models Data Functions
 #' @keywords simStateSpace ssm
 #' @export
-SSMMeanY <- function(nu, lambda, mean_eta) {
-  .SSMMeanY(
-    nu = nu,
-    lambda = lambda,
+SSMInterceptEta <- function(beta, mean_eta) {
+  .SSMInterceptEta(
+    beta = beta,
     mean_eta = mean_eta
   )
 }