+++ R/simStateSpace-ssm-intercept-y.R
@@ -1,26 +1,27 @@
+#' Intercept from
 #' Steady-State Mean Vector for the
 #' Observed Variables in the
 #' State Space Model
 #'
-#' The steady-state mean vector
+#' The intercept vector
 #' for the observed variables
 #' in the state space model
-#' \eqn{\mathrm{Mean} \left( \mathbf{y} \right)}
+#' \eqn{\boldsymbol{\nu}}
 #' is given by
 #' \deqn{
-#'   \mathrm{Mean} \left( \mathbf{y} \right)
-#'   =
 #'   \boldsymbol{\nu}
-#'   +
+#'   =
+#'   \mathrm{Mean} \left( \mathbf{y} \right)
+#'   -
 #'   \boldsymbol{\Lambda}
 #'   \mathrm{Mean} \left( \boldsymbol{\eta} \right)
 #' }
 #' where
-#' \eqn{\boldsymbol{\nu}}
-#' is the vector of intercept values
-#' for the measurement model,
 #' \eqn{\boldsymbol{\Lambda}}
 #' is the matrix of factor loadings,
+#' \eqn{\mathrm{Mean} \left( \mathbf{y} \right)}
+#' is the steady-state mean vector
+#' for the observed variables,
 #' and
 #' \eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}
 #' is the steady-state mean vector
@@ -28,8 +29,15 @@
 #'
 #' @author Ivan Jacob Agaloos Pesigan
 #'
+#' @param mean_y Numeric vector.
+#'   Steady-state mean vector
+#'   of the observed variables
+#'   \eqn{\mathrm{Mean} \left( \mathbf{y} \right)}.
+#' @param mean_eta Numeric vector.
+#'   Steady-state mean vector
+#'   of the latent variables
+#'   \eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}.
 #' @inheritParams SimSSMFixed
-#' @inheritParams SSMInterceptY
 #'
 #' @examples
 #' beta <- matrix(
@@ -47,19 +55,24 @@
 #'   beta = beta,
 #'   alpha = alpha
 #' )
-#' SSMMeanY(
+#' mean_y <- SSMMeanY(
 #'   nu = nu,
 #'   lambda = lambda,
 #'   mean_eta = mean_eta
+#' )
+#' SSMInterceptY(
+#'   mean_y = mean_y,
+#'   mean_eta = mean_eta,
+#'   lambda = lambda
 #' )
 #'
 #' @family Simulation of State Space Models Data Functions
 #' @keywords simStateSpace ssm
 #' @export
-SSMMeanY <- function(nu, lambda, mean_eta) {
-  .SSMMeanY(
-    nu = nu,
-    lambda = lambda,
-    mean_eta = mean_eta
+SSMInterceptY <- function(mean_y, mean_eta, lambda) {
+  .SSMInterceptY(
+    mean_y = mean_y,
+    mean_eta = mean_eta,
+    lambda = lambda
   )
 }