+++ src/source.cpp
@@ -2557,6 +2557,33 @@ arma::mat SSMCovY(const arma::mat& lambda, const arma::mat& theta,
   return ((X + X.t()) / 2);
 }
 // -----------------------------------------------------------------------------
+// edit .setup/cpp/simStateSpace-ssm-intercept-eta-dot.cpp
+// Ivan Jacob Agaloos Pesigan
+// -----------------------------------------------------------------------------
+
+#include <RcppArmadillo.h>
+// [[Rcpp::depends(RcppArmadillo)]]
+// [[Rcpp::export(.SSMInterceptEta)]]
+Rcpp::NumericVector SSMInterceptEta(const arma::mat& beta,
+                                    const arma::vec& mean_eta) {
+  arma::vec output = mean_eta - beta * mean_eta;
+  return Rcpp::NumericVector(output.begin(), output.end());
+}
+// -----------------------------------------------------------------------------
+// edit .setup/cpp/simStateSpace-ssm-intercept-y-dot.cpp
+// Ivan Jacob Agaloos Pesigan
+// -----------------------------------------------------------------------------
+
+#include <RcppArmadillo.h>
+// [[Rcpp::depends(RcppArmadillo)]]
+// [[Rcpp::export(.SSMInterceptY)]]
+Rcpp::NumericVector SSMInterceptY(const arma::vec& mean_y,
+                                  const arma::vec& mean_eta,
+                                  const arma::mat& lambda) {
+  arma::vec output = mean_y + lambda * mean_eta;
+  return Rcpp::NumericVector(output.begin(), output.end());
+}
+// -----------------------------------------------------------------------------
 // edit .setup/cpp/simStateSpace-ssm-mean-eta-dot.cpp
 // Ivan Jacob Agaloos Pesigan
 // -----------------------------------------------------------------------------